refactor: reorganize doctor controllers and update routes for improved structure and clarity
